chore(build): migrate from Plasmo to WXT
Behavior-for-behavior framework swap, no user-facing change and no reset of
existing users' synced settings. Carries the spike's TypeScript 7 + Biome
toolchain into production and removes Plasmo entirely.

- Entrypoints ported to entrypoints/: hn-sort.content (createIntegratedUi
  light-DOM mount + cssInjectionMode 'manifest', reproducing content.tsx),
  comments.content (defineContentScript), background (defineBackground shell
  over app/utils/badge), popup (HTML entrypoint + main.tsx mount)
- Storage stays on @plasmohq/storage (Path 1) — no migration, no reset (R12);
  storage.watch reactivity verified working under WXT (R9)
- Manifest name/description/permissions/host_permissions set in wxt.config.ts;
  icons pre-generated into public/icon/ (no build-time sharp dependency)
- Badge logic extracted to app/utils/badge.ts so its unit tests stay green
- Scripts flipped to wxt/wxt build/wxt zip; tsconfig extends .wxt/tsconfig.json;
  screenshot/demo generators point at .output/chrome-mv3 content-scripts/
- Docs, changelog, description.txt updated; version 2.6.1 -> 2.6.2

Gates: wxt build, tsc --noEmit (TS 7), biome+prettier lint, 251 unit tests,
integration selectors, and both generators all pass. Runtime-smoked in a real
extension context (sort panel, comments, popup, badge).
