+++ .prettierrc.mjs
@@ -14,8 +14,6 @@ export default {
     '<BUILTIN_MODULES>', // Node.js built-in modules
     '<THIRD_PARTY_MODULES>', // Imports not matched by other special words or groups.
     '', // Empty line
-    '^@plasmo/(.*)$',
-    '',
     '^@plasmohq/(.*)$',
     '',
     '^~(.*)$',
