+++ app/background.test.ts
@@ -1,6 +1,7 @@
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
 import { SETTINGS_KEYS } from '~app/constants';
+import { initBadge } from '~app/utils/badge';
 
 const setBadgeText = vi.fn();
 const setBadgeBackgroundColor = vi.fn();
@@ -32,7 +33,7 @@ describe('background service worker', () => {
       cb({ [SETTINGS_KEYS.LAYOUT_OK]: false });
     });
 
-    await import('~/background');
+    initBadge();
 
     expect(setBadgeText).toHaveBeenCalledWith({ text: ':(' });
     expect(setBadgeBackgroundColor).toHaveBeenCalledWith({ color: '#E05050' });
@@ -43,7 +44,7 @@ describe('background service worker', () => {
       cb({ [SETTINGS_KEYS.LAYOUT_OK]: true });
     });
 
-    await import('~/background');
+    initBadge();
 
     expect(setBadgeText).not.toHaveBeenCalled();
   });
@@ -53,7 +54,7 @@ describe('background service worker', () => {
       cb({ [SETTINGS_KEYS.LAYOUT_OK]: true });
     });
 
-    await import('~/background');
+    initBadge();
 
     storageChangeListener({ [SETTINGS_KEYS.LAYOUT_OK]: { newValue: false } });
 
@@ -65,7 +66,7 @@ describe('background service worker', () => {
       cb({ [SETTINGS_KEYS.LAYOUT_OK]: false });
     });
 
-    await import('~/background');
+    initBadge();
     vi.clearAllMocks();
 
     storageChangeListener({ [SETTINGS_KEYS.LAYOUT_OK]: { newValue: true } });
