+++ biome.json
@@ -2,7 +2,7 @@
   "$schema": "https://biomejs.dev/schemas/2.5.3/schema.json",
   "vcs": { "enabled": true, "clientKind": "git", "useIgnoreFile": true },
   "files": {
-    "includes": ["**", "!build", "!.plasmo", "!.wxt", "!.output", "!.claude", "!app/__fixtures__"]
+    "includes": ["**", "!.wxt", "!.output", "!.claude", "!app/__fixtures__"]
   },
   "formatter": { "enabled": false },
   "assist": { "enabled": false },
