--- content.css
@@ -1,258 +1,0 @@
-#hns-control-panel {
-  display: inline-block;
-  position: relative;
-  color: #222;
-  line-height: 12px;
-}
-
-.hns-sort-by-label {
-  display: inline-block;
-  margin-right: 5px;
-  color: #000;
-}
-
-.hns-btn {
-  all: unset;
-  cursor: pointer;
-  display: inline-block;
-  margin: 0 5px;
-}
-
-.hns-btn:focus-visible {
-  outline: 2px solid #fff;
-  outline-offset: 2px;
-  border-radius: 2px;
-}
-
-.hns-btn-text {
-  display: none;
-}
-
-.hns-btn-shortcut {
-  display: inline;
-}
-
-.hns-active {
-  color: #fff;
-}
-
-.hns-btn:hover {
-  opacity: 0.7;
-}
-
-.hns-active.hns-btn {
-  cursor: default;
-}
-
-.hns-active.hns-btn:hover {
-  opacity: 1;
-}
-
-.hns-divider {
-  display: inline-block;
-  margin: 0 10px;
-}
-
-.hns-highlight {
-  font-weight: bold;
-}
-
-.hns-conflict-note {
-  margin-left: 10px;
-  font-size: 11px;
-  color: #828282;
-}
-
-/* Dropdown tier — hidden until the count-aware collapse breakpoints below */
-.hns-dropdown-tier {
-  display: none;
-}
-
-.hns-dropdown {
-  font: inherit;
-  color: #000;
-  /* A native <select> sizes to its widest option ("comments"/"velocity" ~92px), leaving a loose
-   * gap after short values like "points". Cap it so short values sit snug; the two long words clip
-   * by a hair — Chrome keeps the arrow and clips the text, and it's still clear what the control is. */
-  max-width: 84px;
-}
-
-/*
- * Tier 1↔2 (full names ↔ letters). The switch point is count-aware: more options
- * need more width before full names fit, so each enabled-option count (4–6) gets its
- * own breakpoint keyed off the data-sort-count attribute on the panel root (KTD-3).
- * Values calibrated against HN's header; 1280px is the original 4-option baseline.
- */
-@media screen and (min-width: 1280px) {
-  #hns-control-panel[data-sort-count='4'] .hns-btn-text {
-    display: inline;
-  }
-
-  #hns-control-panel[data-sort-count='4'] .hns-btn-shortcut {
-    display: none;
-  }
-}
-
-@media screen and (min-width: 1360px) {
-  #hns-control-panel[data-sort-count='5'] .hns-btn-text {
-    display: inline;
-  }
-
-  #hns-control-panel[data-sort-count='5'] .hns-btn-shortcut {
-    display: none;
-  }
-}
-
-@media screen and (min-width: 1440px) {
-  #hns-control-panel[data-sort-count='6'] .hns-btn-text {
-    display: inline;
-  }
-
-  #hns-control-panel[data-sort-count='6'] .hns-btn-shortcut {
-    display: none;
-  }
-}
-
-/*
- * Tier 2↔3 (letters ↔ dropdown). Also count-aware: the single-letter buttons row is still wide
- * enough to push HN's own nav onto a second line once the viewport is narrow, so we collapse to a
- * compact native <select> *before* that happens. Each breakpoint sits just above the width at which
- * the letters row starts wrapping HN's nav for that option count (measured against HN's header:
- * ~1044/1084/1124px for 4/5/6 options — +margin below). The label-less dropdown is then as narrow as
- * the header allows, so it never wraps the nav until HN's nav would wrap on its own (~750px).
- */
-@media screen and (max-width: 1080px) {
-  #hns-control-panel[data-sort-count='4'] .hns-buttons-tier {
-    display: none;
-  }
-
-  #hns-control-panel[data-sort-count='4'] .hns-dropdown-tier {
-    display: inline-block;
-  }
-}
-
-@media screen and (max-width: 1120px) {
-  #hns-control-panel[data-sort-count='5'] .hns-buttons-tier {
-    display: none;
-  }
-
-  #hns-control-panel[data-sort-count='5'] .hns-dropdown-tier {
-    display: inline-block;
-  }
-}
-
-@media screen and (max-width: 1160px) {
-  #hns-control-panel[data-sort-count='6'] .hns-buttons-tier {
-    display: none;
-  }
-
-  #hns-control-panel[data-sort-count='6'] .hns-dropdown-tier {
-    display: inline-block;
-  }
-}
-
-/*
- * Below ~750px HN's header wraps to multiple lines — the login/username drops to its own line
- * beneath our <select>, which strands the trailing "|" at the end of a line with nothing after it.
- * Drop the separator there (it's only meaningful while the select and the header links share a line).
- */
-@media screen and (max-width: 750px) {
-  .hns-dropdown-tier .hns-divider {
-    display: none;
-  }
-}
-
-/* New post indicator - only active when table body has .hns-show-new */
-.hns-show-new tr.hns-new-post .titleline::before {
-  content: '\2022';
-  color: #ff6600;
-  margin-right: 4px;
-  font-weight: bold;
-  opacity: var(--hns-fade, 1);
-}
-
-/* Review toast (speech bubble below sort menu) */
-.hns-review-toast {
-  position: absolute;
-  top: 22px;
-  left: 10px;
-  display: inline-flex;
-  align-items: flex-start;
-  gap: 8px;
-  background: #fff;
-  border: 1px solid #ddd;
-  border-radius: 8px;
-  padding: 8px 24px 8px 10px;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.12);
-  font-size: 12px;
-  color: #444;
-  white-space: nowrap;
-  z-index: 1000;
-  text-align: left;
-}
-
-/* Arrow pointing up toward "sort by:" */
-.hns-review-toast::before {
-  content: '';
-  position: absolute;
-  top: -7px;
-  left: 20px;
-  width: 12px;
-  height: 12px;
-  background: #fff;
-  border-left: 1px solid #ddd;
-  border-top: 1px solid #ddd;
-  transform: rotate(45deg);
-}
-
-.hns-review-link {
-  display: block;
-  color: #444;
-  font-weight: normal;
-  text-decoration: none;
-  text-align: left;
-  transition: color 0.15s;
-}
-
-.hns-review-link:hover {
-  color: #ff6600;
-  text-decoration: none;
-}
-
-.hns-review-sub {
-  display: block;
-  font-size: 11px;
-  color: #999;
-  margin-top: 4px;
-}
-
-.hns-review-close {
-  all: unset;
-  cursor: pointer;
-  position: absolute;
-  top: 4px;
-  right: 6px;
-  color: #bbb;
-  font-size: 18px;
-  line-height: 1;
-}
-
-.hns-review-close:hover {
-  color: #888;
-}
-
-.hns-review-close:focus-visible {
-  outline: 2px solid #ff6600;
-  outline-offset: 2px;
-  border-radius: 2px;
-}
-
-@media (prefers-reduced-motion: reduce) {
-  .hns-review-link {
-    transition: none;
-  }
-
-  .hns-show-new tr.hns-new-post .titleline::before {
-    opacity: 1;
-  }
-}
