+++ package.json
@@ -1,18 +1,15 @@
 {
   "name": "hacked_news_sorted",
-  "displayName": "Hacker News Sorted",
-  "version": "2.6.1",
+  "version": "2.6.2",
   "description": "Instantly sort Hacker News by points, time, comments, velocity, or heat, mark new posts, and highlight comment authors.",
   "author": "Leonid Svyatov <leonid@svyatov.com>",
   "license": "MIT",
   "type": "module",
   "scripts": {
-    "dev": "plasmo dev",
-    "build": "plasmo build",
-    "package": "plasmo package",
+    "dev": "wxt",
+    "build": "wxt build",
+    "package": "wxt zip",
     "release": "bun run build && bun run package",
-    "spike:build": "wxt build",
-    "spike:prepare": "wxt prepare",
     "test": "vitest run",
     "test:integration": "vitest run app/utils/selectors.integration.test.ts app/utils/comments.integration.test.ts",
     "test:watch": "vitest",
@@ -21,6 +18,7 @@
     "fixture:update": "bun app/__fixtures__/updateFixture.ts",
     "screenshots": "bun scripts/generate-screenshots.ts",
     "demo": "bun scripts/generate-demo.ts",
+    "postinstall": "wxt prepare",
     "prepare": "simple-git-hooks"
   },
   "simple-git-hooks": {
@@ -34,7 +32,6 @@
     "@plasmohq/storage": "^1.15.0",
     "baseline-browser-mapping": "^2.10.43",
     "caniuse-lite": "^1.0.30001799",
-    "plasmo": "0.90.5",
     "react": "^19.2.7",
     "react-dom": "^19.2.7"
   },
